+++ package.json
@@ -12,7 +12,7 @@
   },
   "dependencies": {
     "@supabase/supabase-js": "^2.47.10",
-    "@tanstack/react-query": "^5.62.8",
+    "@tanstack/react-query": "^5.62.10",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-error-boundary": "^5.0.0",
