--- src/components/common-ui/header/Header.styles.ts
@@ -1,7 +1,0 @@
-import styled from 'styled-components'
-
-export const ProfileImg = styled.img`
-  border-radius: 50%; // 디자인 토큰으로 변경 필요
-  width: 50px;
-  height: 50px;
-`
