+++ src/components/common-ui/header/header.styled.ts
