+++ src/components/common-ui/star-rating/StarRating.styled.ts
