+++ src/components/common-ui/star-rating/StarRating.tsx
