+++ src/layout/DefaultLayout.tsx
@@ -1,10 +1,10 @@
 import { Outlet } from 'react-router-dom'
-import Header from '@/components/header/header'
+import Header from '@/components/common-ui/header/Header'
 
 export default function DefaultLayout() {
   return (
     <>
-      <Header />
+      <Header backgroundColor="black" />
       <Outlet />
     </>
   )
